requirements.txt
NorenRestApiPy==0.0.22
pyotp==2.9.0
pandas==3.0.1
requests==2.32.5
websocket-client==1.9.0
Flask==3.0.2
phidata
duckdb
openpyxl
streamlit
